Pushing the current working code
`parsers.py` contains the class definitions for the various dataset types (Roster, Timesheet, Incidents). 
`character.py` contains the class definition for a character, which is used to house all the data associated with a specific character. 
`parsePDCSVs.py` is an example script that I used to produce the data shown on THE SHEET. 
`__init__.py` just is.