examples of check_names function succeeding and failing
The `check_names` function in the `parsers.py` file uses a hard-set limit of string similarity for names to be mapped to other names. I have some print statements about when this mapping happens and when it maybe should be happening. This file is the unique set of messages associated with this.